+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Miscellaneous
 - Add tests
     - positive/negative tests for action
+    - positive/negative tests for action type
 
 ## [1.0.1] - 2018-02-10